pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synthtiger"
version = "0.1.0"  # You may want to use a dynamic version from _version.py
description = "Synthetic text image generator for OCR model"
readme = "README.md"
authors = [
    {name = "Moonbin Yim"},
    {name = "Yoonsik Kim"},
    {name = "Han-Cheol Cho"},
    {name = "Sungrae Park"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.6"
dependencies = [
    "arabic-reshaper==2.1.3",
    "blend-modes==2.1.0",
    "fonttools==4.33.3",
    "imgaug==0.4.0",
    "numpy==1.22.4",
    "opencv-python==4.6.0.66",
    "pillow>=8.2.0",
    "pygame==2.1.2",
    "python-bidi==0.4.2",
    "pytweening==1.0.4",
    "pyyaml==6.0",
    "regex==2022.6.2",
    "scipy==1.8.1",
]

[project.urls]
Homepage = "https://github.com/clovaai/synthtiger"

[project.scripts]
synthtiger = "synthtiger.main:main"

[tool.setuptools]
packages = {find = {exclude = ["tests"]}}
include-package-data = true

[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38', 'py39', 'py310']

[tool.isort]
profile = "black"